[feat] #86 에셋삭제
